Rename hidden as model, and separate out gradient utilities into a separate file
