+++ examples/example_mnist_hopfield.py
@@ -1,6 +1,6 @@
 from paysage import batch
 from paysage import layers
-from paysage.models import hidden
+from paysage.models import model
 from paysage import fit
 from paysage import optimizers
 from paysage import backends as be
@@ -30,7 +30,7 @@ def example_mnist_hopfield(paysage_path=None, num_epochs=10, show_plot=False):
     vis_layer = layers.BernoulliLayer(data.ncols)
     hid_layer = layers.GaussianLayer(num_hidden_units)
 
-    rbm = hidden.Model([vis_layer, hid_layer])
+    rbm = model.Model([vis_layer, hid_layer])
     rbm.initialize(data)
 
     # set up the optimizer and the fit method
