+++ paysage/models/__init__.py
@@ -1,6 +1,7 @@
-from .initialize import init_hidden
+from .initialize import init_model
 from .initialize import init_recurrent
 from .initialize import init_deep
-from . import hidden
+from . import model
+from . import tap_machine
 from . import recurrent
 from . import deep
